feat: 공동 구매 참여 메소드 GroupBuyUserService로 이동
